# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Layal Al Zahra Salon CRM API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /staff:
    get:
      operationId: listStaff
      tags: [staff]
      summary: List all staff members
      responses:
        "200":
          description: Staff list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StaffMember"
    post:
      operationId: createStaff
      tags: [staff]
      summary: Create a new staff member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StaffInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffMember"
  /staff/{id}:
    get:
      operationId: getStaff
      tags: [staff]
      summary: Get staff member by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Staff member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffMember"
        "404":
          description: Not found
    patch:
      operationId: updateStaff
      tags: [staff]
      summary: Update a staff member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StaffUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffMember"
    delete:
      operationId: deleteStaff
      tags: [staff]
      summary: Delete a staff member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /staff/{id}/schedule:
    get:
      operationId: getStaffSchedule
      tags: [staff]
      summary: Get working hours for a staff member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Schedule
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WorkingHours"
    put:
      operationId: updateStaffSchedule
      tags: [staff]
      summary: Update working hours for a staff member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkingHoursUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WorkingHours"
  /staff/{id}/performance:
    get:
      operationId: fetchStaffPerformance
      tags: [staff]
      summary: Get staff performance & sales stats
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Performance stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffPerformance"

  # CLIENTS
...
components:
  schemas:
...
    StaffMember:
      type: object
      required: [id, name, email, role, isActive, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        role:
          type: string
          enum: [admin, manager, receptionist, stylist, therapist]
        specialization:
          type: ["string", "null"]
        color:
          type: ["string", "null"]
        commissionRate:
          type: ["number", "null"]
        targetMonthly:
          type: ["number", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
    StaffInput:
      type: object
      required: [name, email, password, role]
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [admin, manager, receptionist, stylist, therapist]
        specialization:
          type: string
        color:
          type: string
        commissionRate:
          type: number
        targetMonthly:
          type: number
    StaffUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [admin, manager, receptionist, stylist, therapist]
        specialization:
          type: string
        color:
          type: string
        commissionRate:
          type: number
        targetMonthly:
          type: number
        isActive:
          type: boolean
    WorkingHours:
      type: object
      required: [id, staffId, dayOfWeek, startTime, endTime, isWorking]
      properties:
        id:
          type: integer
        staffId:
          type: integer
        dayOfWeek:
          type: integer
          description: "0=Sunday, 1=Monday, ..., 6=Saturday"
        startTime:
          type: string
        endTime:
          type: string
        isWorking:
          type: boolean
    WorkingHoursUpdate:
      type: object
      required: [hours]
      properties:
        hours:
          type: array
          items:
            type: object
            required: [dayOfWeek, startTime, endTime, isWorking]
            properties:
              dayOfWeek:
                type: integer
              startTime:
                type: string
              endTime:
                type: string
              isWorking:
                type: boolean
    StaffPerformance:
      type: object
      required: [staffId, totalRevenue, totalAppointments, completedAppointments, commissionEarned, targetProgress]
      properties:
        staffId:
          type: integer
        totalRevenue:
          type: number
        totalAppointments:
          type: integer
        completedAppointments:
          type: integer
        commissionEarned:
          type: number
        targetProgress:
          type: number
        topServices:
          type: array
          items:
            $ref: "#/components/schemas/TopService"

    # Clients
...
    TopService:
      type: object
      required: [serviceId, serviceName, count, revenue]
      properties:
        serviceId:
          type: integer
        serviceName:
          type: string
        count:
          type: integer
        revenue:
          type: number
...